+++ changelog.md
@@ -11,6 +11,7 @@ Codion Change Log
 - Select.having() added, related refactoring.
 - EntityConnection.iterator() added.
 - EntityResultIterator now extends Iterable.
+- EntityQueries.instance() renamed factory().
 ### is.codion.framework.db.local
 - LocalEntityConnection.iterator() removed.
 ### is.codion.framework.db.rmi